requirements.txt
pytest~=6.2.3
PyJWT~=2.0.1
Flask~=1.1.2
Werkzeug~=0.16.1
requests~=2.25.1
setuptools~=54.2.0
Flask-JWT-Extended==4.1.0
Flask-SQLAlchemy==2.5.1
flask-restx==0.3.0
PyMySQL==1.0.2


